Move non-component logic to 'lib' folder.
